Decouple GET and POST Api Test
